Remove useThrottledHoverTooltip; call resolveTooltip directly
The rAF throttle + same-pixel-skip in useThrottledHoverTooltip didn't earn its
keep: deck already coalesces hover picking to one pass per animation frame, and
in the default 'simple' mode tooltip resolution just reuses deck's existing pick
(no extra work to throttle). The hook only added a frame of latency and
indirection.

Drop the hook and call `resolveTooltip` straight from the hover handlers. Keep
the one durable piece — the `isHoverDuringDrag` pointer-button drag-gate — moved
into `featureTooltipHover` (its natural home) alongside the other hover helpers,
with its test folded into featureTooltipHover.spec.
